purchase delivery address

+++ app/Http/Controllers/Purchase/PurchaseInvoiceController.php
@@ -13,9 +13,8 @@ class PurchaseInvoiceController extends Controller
 
     public function index(Request $request)
     {
-        $page = $request->get('page') ?? 1;
-        $limit = $request->get('limit') ?? 10;
-        return PurchaseInvoiceService::index($page, $limit);
+        $requestAll = $request->all();
+        return PurchaseInvoiceService::index($requestAll);
     }
 
     public function create(Request $request)
@@ -24,49 +23,86 @@ class PurchaseInvoiceController extends Controller
             $data = $request->all();
             $validator = Validator::make($data, [
                 'number'=>'required|unique:purchase_invoices',
-                'date'=>'required',
-                'summ'=>'required',
                 'purchase_id' => 'required',
             ]);
  
             if($validator->fails()){
                 $error = $validator->errors()->toArray();
-               return response()->json(['message'=>$error])->setStatusCode(417); 
+                return response()->json(['message'=>$error])->setStatusCode(417); 
             
             }
-
             return PurchaseInvoiceService::create($data);
         } catch (Exception $e){
             return $e->getMessage();
         }
     }
 
-     public function card(string $id)
+    public function card(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+                'id'=>'required',
+        ]);
+        if($validator->fails()){
+            $error = $validator->errors()->toArray();
+            return response()->json(['message'=>$error])->setStatusCode(417);     
+        }
+        $id = $request->input('id');
         $data = PurchaseInvoiceService::card($id);
         if(!$data) return response()->json(['message'=>'Not found'], 404);
         return $data; 
     }
 
-    public function update(Request $request, string $id)
+     public function update(Request $request)
     {
-        $data = $request->all();
-        $result = PurchaseInvoiceService::update($id, $data);
-        if(!$result) return response()->json(['message'=>'Not found'], 404);
-        return $result;
+        try {
+            $requestData=$request->all();
+            $validator = Validator::make($requestData, [
+                'id'=>'required',
+                'data'=>'required',
+            ]);
+            if($validator->fails()){
+                $error = $validator->errors()->toArray();
+                return response()->json(['message'=>$error])->setStatusCode(417); 
+            
+            }
+            $id = $request->input('id');
+            $data = $request->input('data');
+            $result = PurchaseInvoiceService::update($id, $data);
+            if(!$result) return response()->json(['message'=>'Not found'], 404);
+            return $result;
+        } catch (Exception $e){
+            return $e->getMessage();
+        }
     }
 
-    public function destroy(string $id)
+    public function destroy(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'id'=>'required',
+        ]);
+        if($validator->fails()){
+            $error = $validator->errors()->toArray();
+            return response()->json(['message'=>$error])->setStatusCode(417); 
+        }
+        $id = $request->input('id');
         $data = PurchaseInvoiceService::delete($id);
         if(!$data) return response()->json(['message'=>'Not found'], 404);
         return $data;
     }
 
-    public function recover(string $id)
+    public function recover(Request $request)
     {
+        $validator = Validator::make($request->all(), [
+            'id'=>'required',
+        ]);
+        if($validator->fails()){
+            $error = $validator->errors()->toArray();
+            return response()->json(['message'=>$error])->setStatusCode(417); 
+        }
+        $id = $request->input('id');
         $data = PurchaseInvoiceService::recover($id);
         if(!$data) return response()->json(['message'=>'Not found'], 404);
         return $data;
     }
+
 }
